+++ include/sphere_simulation.h
@@ -7,14 +7,14 @@
 #include "sphere.h"
 
 struct Event {
-  double time;
+  long double time;
   sphere *s1;
   sphere *s2;
 
   bool operator<(const Event &other) const {
     return time > other.time; // for min-heap
   }
-  Event(double time, sphere *s1, sphere *s2) : time(time), s1(s1), s2(s2) {}
+  Event(long double time, sphere *s1, sphere *s2) : time(time), s1(s1), s2(s2) {}
 };
 
 class sphere_simulation {
@@ -30,7 +30,7 @@ public:
 
   inline sphere *get_spheres() const { return spheres; }
   inline int get_number_of_spheres() const { return number_of_spheres; }
-  inline const std::vector<double> get_collision_times() const {
+  inline const std::vector<long double> get_collision_times() const {
     return collision_times;
   }
 
@@ -45,13 +45,13 @@ public:
 
 private:
   // data structs
-  std::vector<double> collision_times; // stores the collision times
+  std::vector<long double> collision_times; // stores the collision times
   std::priority_queue<Event> event_queue;
   sphere *spheres;
 
   // simulation parameters
-  double max_time;
-  double current_time;
+  long double max_time;
+  long double current_time;
   int number_of_spheres;
   double torus_size;
   double epsilon; // radius
@@ -61,7 +61,7 @@ private:
   void add_collision_event(sphere *s1, sphere *s2);
   void find_collision_events(sphere *s1);
   void wrap_around(sphere *s);
-  void update_positions(double dt);
+  void update_positions(long double dt);
   std::vector<point3> get_images(sphere *s);
 };
